requirements.txt
pydantic>=2.0.0
pytest>=7.0.0
mypy>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
openai>=1.0.0
anthropic>=0.7.0 